Juego terminado
Falta la documentacion

+++ MonsterTECG/src/main/java/monsterTecg/Logics/DesignPatterns/Template/MixFacade.java
@@ -12,18 +12,20 @@ public class MixFacade extends FacadePerformer{
 
     @Override
     public void makeChanges(int dmg) {
-        while (PlayerManager.getInstance().getHand().getSize() < 10){
-            PlayerManager.getInstance().getHandUpdate();
+        int currentCards = PlayerManager.getInstance().getHand().getSize();
+        if (currentCards + PlayerManager.getInstance().deckSize >= 10){
+            while (PlayerManager.getInstance().getHand().getSize() < 10){
+                PlayerManager.getInstance().getHandUpdate();
+            }
+
+            while (PlayerManager.getInstance().getHand().getSize() != currentCards){
+                Random rand = new Random();
+                int rnd = rand.nextInt(PlayerManager.getInstance().getHand().getSize());
+                Card current = PlayerManager.getInstance().getHand().getByIndex(rnd);
+                PlayerManager.getInstance().getDeck().push(current);
+                PlayerManager.getInstance().getHand().deleteByIndex(rnd);
+            }
         }
-        
-        while (PlayerManager.getInstance().getHand().getSize() != 4){
-            Random rand = new Random();
-            int rnd = rand.nextInt(PlayerManager.getInstance().getHand().getSize());
-            Card current = PlayerManager.getInstance().getHand().getByIndex(rnd);
-            PlayerManager.getInstance().getDeck().push(current);
-            PlayerManager.getInstance().getHand().deleteByIndex(rnd);
-        }
-        
         
     }
